Add files via upload
These are some incomplete sections for the message that will be sent by the nodes to try and determine if every node has received the update. 3D3Project2.cpp contains some code that tests the classes and a rough organisation of what the code running on the nodes might look like for that section. I still have to write the message that will be sent by the nodes when they detect that all nodes have the updated list. Finally we'll have to decide on a final structure for the header files so that all of our code works together.